+++ zephyr/src/coredumps/spotflow_coredumps_cbor.h
@@ -9,6 +9,9 @@
 extern "C" {
 #endif
 
+/* Should be approximately 57 bytes (including build ID), putting 64 to be safe */
+#define SPOTFLOW_COREDUMPS_CBOR_OVERHEAD 64
+
 int spotflow_cbor_encode_coredump(const uint8_t* coredump_data, size_t coredump_data_len,
 				  int chunk_ordinal, uint32_t core_dump_id, bool last_chunk,
 				  const uint8_t* build_id, size_t build_id_len,